+++ src/test/java/com/farmdora/farmdora/ControllerTest.java
@@ -7,6 +7,7 @@ import com.farmdora.farmdora.opinion.service.OpinionService;
 import com.farmdora.farmdora.sale.controller.SaleController;
 import com.farmdora.farmdora.sale.controller.SellerSaleController;
 import com.farmdora.farmdora.sale.service.SaleService;
+import com.farmdora.farmdora.sale.service.SellerSaleService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
@@ -29,6 +30,9 @@ public abstract class ControllerTest {
 
     @MockitoBean
     protected SaleService saleService;
+
+    @MockitoBean
+    protected SellerSaleService sellerSaleService;
 
     @MockitoBean
     protected OpinionService opinionService;
